trigger check updates
calls all core functions to check for updates as soon as you enter the Debug Information tab.

I made this as an extra function just in case we want to call it from elsewhere in the future as well.

Also changed some capitalizations in strings to be more consistent as it was a mix-match at the moment.

( fixes #67 )