close diff view with esc
adds a check for esc button to close the diff view and clear the window as well.

also I noticed that the default 5ms on slow connections is giving some problems as the checksum list is a bit long and sometimes hangs when waiting for a reply, so I bumped it to 10ms instead to see if that's better.

( fixes #76 )