Fix plugin actions links in plugin_actions() and troubleshoot_plugin_action() functions
$plugin_slug is correctly set in health_check_troubleshoot_menu_bar() and also in display_dashboard_widget() based on slugs stored in $this->active_plugins

plugin_actions() and troubleshoot_plugin_action() functions are using a different approach using $plugin_file for plugins living in wp-content/plugins/ root (this is correct), but incorrently assuming $plugin_data['Name'] it's always the same than the plugin slug for any plugin living in a folder (most common scenario).

This commit fixes the issue using $plugin_file to get the correct value for $plugin_slug also for plugins living in their own folder, the most common scenario. That is basically the same that the other functions do using $this->active_plugins