Troubleshooting Mode: Introduce more widespread nonce use.
Nonces were initially omitted from various actions by design to ease the process of support volunteers directing users, specifically as it would only be a simulated environment tied to a single user session.

Despite this, code scanners will still remark on the lack of validation, so introducing nonces will remove unnecessary worry for users.

This implements a user-oriented prompt if a nonce is either missing, or invalid, giving them an explanation of what is happening, what action is attempted, and clear paths to either approve, or cancel, this behavior.