Tools: Add unexpected file detection to integrity checker.
Validation of core checksums is nice, but if there are other unexpected files in the filesystem, the integrity of them is _technically_ not passing.

This introduces a second pass of any files in the core file locations that will alert of unknown files (files that do not exist within the checksum response).

Props @Zodiac1978.
Fixes #259.