Tools: Check the time it takes to send an e-mail with `wp_mail`.
E-mail deliverables are complicated things, and although `wp_mail` may not be triggering errors, if a SMTP plugin is in use for example, a prolonged sending time may indicate upstream issues, or an overloaded local system in cases where it is used.

This will warn users if an email message took more than 3 seconds to send.

Fixes #376.