Styles: Improve some of the styling shared between the plugin and core.
Core has added various styles since the Site Health module was incorporated, and although that element is no longer part of the plugin, the styles are still used elsewhere in the plugin and led to some less than ideal user experiences on certain breakpoints and smaller devices.

This pulls in applicable core styles to the plugin to ensure a consistent experience when the plugin is enabled.