Screenshots: Introduce the new screenshot beta feature.
Screenshots are often vital to a support request, but not everyone is familiar with how to take one, or how to share it. This introduces a beta feature with a fixed button that will both take a screenshot, and store it on your site (outside the media library), and then provides the user both the ability to view them, copy the image code for sharing directly on the support forums, and even own the content by being able to delete it as they wish.

The feature is currently in beta mode, as screenshots may not always accurately represent the markup of a site, and is there to gather wider feedback on implementation and behavior.